+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 *   Add metric filters
 *   Remove unused variables
+*   Add metric alarms
 
 ### 0.0.3